fix: remote config polling parsing (#123)
According to the Remote Configuration specification,
`DD_REMOTE_CONFIG_POLL_INTERVAL_SECONDS` support floating point input
and can be set to `0`.

+++ src/datadog/datadog_agent_config.h
@@ -61,7 +61,7 @@ struct DatadogAgentConfig {
   Optional<bool> remote_configuration_enabled;
   // How often, in seconds, to query the Datadog Agent for remote configuration
   // updates.
-  Optional<int> remote_configuration_poll_interval_seconds;
+  Optional<double> remote_configuration_poll_interval_seconds;
 
   static Expected<HTTPClient::URL> parse(StringView);
 };
